separate out middleware
